shopping_cart/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.11.0
pipenv==2018.11.26
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
